✨ Refactor: Extract components and add Jest testing
🏗️ COMPONENT ARCHITECTURE:
- Extract MemberCard component for individual member display
- Extract DatabaseStatus component for connection indicator
- Extract MemberGrid component for layout + animations
- Clean up monolithic page.tsx into focused components

🧪 TESTING SETUP:
- Add Jest + React Testing Library + @types/jest
- Configure Next.js Jest integration with moduleNameMapper
- Add basic component tests for DatabaseStatus and MemberCard
- Add test scripts: yarn test and yarn test:watch

🎯 DEMO-READY:
- Much better structure for showing AI Meetup devs
- Separation of concerns with reusable components
- Professional testing setup demonstrates best practices
- Maintains all existing functionality and animations

Ready for prime time! 🚀
